refactor(io): 불필요한 개행 제거 및 출력방식 변경
ConsoleDevice 출력방식: sout -> terminal
