+++ src/main/java/org/devcourse/springbasic/io/ConsoleDevice.java
@@ -2,21 +2,21 @@ package org.devcourse.springbasic.io;
 
 import org.beryx.textio.TextIO;
 import org.beryx.textio.TextIoFactory;
+import org.beryx.textio.TextTerminal;
 import org.devcourse.springbasic.validator.Validator;
 import org.devcourse.springbasic.validator.VoucherMenuValidator;
-import org.devcourse.springbasic.menu.MenuType;
+import org.devcourse.springbasic.controller.menu.MenuType;
 import org.devcourse.springbasic.validator.MenuValidator;
 import org.devcourse.springbasic.voucher.Voucher;
 import org.devcourse.springbasic.voucher.VoucherType;
-import org.springframework.stereotype.Component;
 
 import java.text.MessageFormat;
 import java.util.List;
 
-@Component
 public class ConsoleDevice implements IODevice{
 
     private final TextIO textIO = TextIoFactory.getTextIO();
+    private final TextTerminal<?> terminal = TextIoFactory.getTextIO().getTextTerminal();
 
     @Override
     public MenuType inputMenu() {
@@ -26,7 +26,6 @@ public class ConsoleDevice implements IODevice{
 
         Validator<String> menuValidator = new MenuValidator<>();
         boolean isValid = menuValidator.validate(input);
-
         if (isValid) {
             return MenuType.findSolveStateByInput(input);
         } else {
@@ -47,16 +46,15 @@ public class ConsoleDevice implements IODevice{
         } else {
             throw new IllegalArgumentException("잘못된 메뉴 입력입니다.");
         }
-
     }
 
     @Override
     public void outputVoucher(Voucher voucher) {
-        System.out.println(MessageFormat.format("생성된 바우처 정보입니다 => {0}", voucher));
+        terminal.printf(MessageFormat.format("생성된 바우처 정보입니다 => {0}", voucher));
     }
 
     @Override
     public void outputVouchers(List<Voucher> vouchers) {
-        System.out.println(MessageFormat.format("바우처 목록입니다\n {0}", vouchers));
+        terminal.printf(MessageFormat.format("바우처 목록입니다\n {0}", vouchers));
     }
 }
